renamed ssl_message_translator to ssl_geometry_message_translator
