+++ src/software/proto/message_translation/ssl_geometry_message_translator.h
